Changes :
- Added manual announce option
- Torrent::getHash now throws an error if idx is out of bounds
- Get rid of some signed/unsigned mismatch warnings
- Implemented 109247 : Download complete, seeding -> Seeding
- Implemented 109249 : limit length of file field in KTorrentView

FEATURE: 109247
FEATURE: 109249
GUI


svn path=/trunk/extragear/network/ktorrent/; revision=435823

+++ libtorrent/udptracker.cpp
@@ -247,7 +247,7 @@ namespace bt
 			peer_buf = new Uint8[ba];
 		}
 		
-		if (sock->readBlock((char*)peer_buf,ba) != ba)
+		if ((Uint32)sock->readBlock((char*)peer_buf,ba) != ba)
 		{
 			handleError();
 			return;
